Created routes and services and controller logic for todolist Added custom exceptions Changed project structure by moving directories from HTTP folder, Changed delete function in model by requirements  Added function for checking todolist status Added method gor serching todolist by task id in task repository

+++ app/Services/TodoListService.php
@@ -55,8 +55,14 @@ class TodoListService
         return $this->repository->update($todoList, $data);
     }
 
-    public function delete(TodoList $todoList): bool
+    public function delete(int $id): bool
     {
+        $todoList = $this->repository->getById($id);
+
+        if (!$todoList) {
+            throw new TodoListNotFoundException('TodoList not found');
+        }
+
         return $this->repository->delete($todoList);
     }
